lruttl
Struct LruCacheWithTtl

pub struct LruCacheWithTtl<K: Clone + Hash + Eq, V: Clone> { /* private fields */ }

Implementations

impl<K: Clone + Hash + Eq + Send + 'static, V: Clone + Send + 'static> LruCacheWithTtl<K, V>

pub fn new(capacity: usize) -> Self

👎Deprecated: use new_named instead

pub fn new_named<S: Into<String>>(name: S, capacity: usize) -> Self

pub fn clear(&self) -> usize

pub fn get_with_expiry<Q>(&self, name: &Q) -> Option<(V, Instant)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

pub fn get<Q>(&self, name: &Q) -> Option<V>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

pub fn insert(&self, name: K, item: V, expiration: Instant) -> V

pub fn prune_expired(&self) -> usize

pub fn get_or_insert<F: FnOnce() -> V>( &self, name: K, ttl: Duration, func: F, ) -> V

Get an existing item, but if that item doesn’t already exist, call func to provide a value that will be inserted and then returned. This is done atomically wrt. other callers.